add country leaderboard
Added the api endpoint on server.js as well as the front end react code for it along with the button required for users to view it. Also added some navigation such that when users log in, they don't remain on that page but go back to the main page to play the game (while logged in). same goes for after they submit a game.